We should be doing all content-based validation server-side
This removes content based validation, but leaves in place
type based validation.  The motivation here is that the presense
of LinkByUIDs makes validation of un-liked objects incomplete
and therefore partially inconsistent.  Validation should be the
responsibility of the data store, not the data objects themselves.

At Citrine, this means that we implement content-based validation
server side, which let's us include validations that traverse links
between the different data objects.  If we support client-side validation,
it will be through making a call to our data service.