Use seen instead of skip to break cycles in recursive_foreach
The seen set of object hashes should be enough to avoid cycles,
so we don't have to re-use the skip member from DictSerializable
(which conflated serialization and pointer concerns).

Also moved the applyFirst/applyLast logic out of those if statements
to prevent the first object in a cycle from having the function applied
to it twice.