Remove type hints because we don't strip them during deployment
We were using strip-hints to provide python 3.5 support.  However,
those hints aren't being stripped in the distribution that we
publish to pypi.  Rather than try to roll forward, we're rolling
back type hints for now to maintain the ability to test against
python 3.5 in citrine-python.

These will be added back when we have a process for stripping
them for deployments.